improve power management. serial Tx turned into open drain.
Turning serial TX into an open drain may cause some compatibility
problems, as the board currently does not have a pullup resistor on
it. I think the FTD232R has a pull-up which is why mine is working
but there is a possibility that it's just parasitic. Anyways, the
impact may be that serial characters on the screen could be garbled.

This fix patches an issue where if it's an active-drive, the active
drive actually powers the serial chip sufficiently to force it into
reset which then forces the CPU reset pin, causing a reset loop where
the CPU and serial adapter successively take turns resetting each other.