add battery voltage measurement routine. 'v' will print a battery code
The battery code is a value from 0 - BATT_RANGE. 0 means you're out of
juice, and the system should shut down to prevent battery damage.
The top value means the battery is fully charged. At the moment,
BATT_RANGE is set to 16, which allows you to do, for example, a 16-pixel
battery meter in the UI.

The accuracy of the battery reporting is as good as +/-3.4%, based upon
the limitation of the VREF calibration source inside the chip. It should
actually be probably tighter than +/-2% in practice because the broader
range is over an extended temperature range from -40C to +85C, and one
of the biggest factors affecting the accuracy of the reference will be
temperature.