safecast: Major refactor of code
This moves everything to a driver-based model wherein each device
attached to the system has an init, resume, suspend, and deinit
routine.

When devices are first added, their "init" routine is called.  They
start out in the "SUSPENDED" state and must be "resumed".  Various
power events can cause devices to be "paused" or "removed".

Devices are paused and removed in the reverse order from when they
were added.  Thus, it is a good idea to, for example, add the "power"
device first so that it gets shut down last.

The "switch" device (which represents the switch on the back)
currently makes calls to the power management system to pause
or resume everything whenever the switch is hit.