modifications to power code
wfe/wfi routines tweaked
wfe/wfi call now appended to the end of the actual calls
started debugging why wfe/wfi doesn't work

added extra code around the battery measerument. not sure if it's
all necessary, as i guess the device won't be waking up from
standby anymore to do the battery measurement. probably will need
to strip a lot of the re-initialization out to reduce the overhead
on battery measurement? although, we could do clock gating to the ADC
now that this code is here and save a lot of power. That may be
the right way to go.